Update Server - Orders

+++ backend/src/controllers/Orders/removeOrder.js
@@ -1,6 +1,15 @@
 const { Order } = require("../../db");
+const findOrderbyId = require("./findOrderbyId");
+
 
 const removeOrder = async (id) => {
+  const order = await findOrderbyId(id);
+  if (!order) {
+    return {
+      message: `No existe una Orden con el id '${id}' para eliminar`,
+    };
+  }
+
   try {
     orderDestroyed = await Order.destroy({
       where: { id: id },
